Removing data to another module
